+++ idscp2/src/main/java/de/fhg/aisec/ids/idscp2/idscp_core/finite_state_machine/StateWaitForRatVerifier.java
@@ -21,7 +21,7 @@ public class StateWaitForRatVerifier extends State {
                                    Timer ratTimer,
                                    Timer handshakeTimer,
                                    Timer verifierHandshakeTimer,
-                                   int ratTimeoutDelay) {
+                                   long ratTimeoutDelay) {
 
         /*---------------------------------------------------
          * STATE_WAIT_FOR_RAT_VERIFIER - Transition Description
